formatting

+++ levelData.cpp
@@ -100,13 +100,13 @@ bool LevelData::inputData() {
 		for (int i = 0; i < gridHeight; i++, x = -1.f, y += gridElementHeight) {
 			for (int j = 0; j < gridWidth; j++, x += gridElementWidth) {
 				//top left
-				gridElement[std::make_pair(i, j)].push_back({x, y + (float)(gridElementHeight)});
+				gridElement[std::make_pair(i, j)].push_back({x, y + gridElementHeight});
 				//bottom left
 				gridElement[std::make_pair(i, j)].push_back({x, y});
 				//bottom right
-				gridElement[std::make_pair(i, j)].push_back({x + (float)(gridElementWidth), y});
+				gridElement[std::make_pair(i, j)].push_back({x + gridElementWidth, y});
 				//top right
-				gridElement[std::make_pair(i, j)].push_back({x + (float)(gridElementWidth), y + (float)(gridElementHeight)});
+				gridElement[std::make_pair(i, j)].push_back({x + gridElementWidth, y + gridElementHeight});
 			}
 		}
 		return true;
